voice-live-on-mobile/hosted-agent2/requirements.txt
azure-ai-agentserver-agentframework==1.0.0b16
httpx
debugpy
# Agent development CLI tool (preview)
agent-dev-cli